Add database configuration socketTimeout
The configuration will be applied to the underlying datasource / driver.
Postgres and MySQL for example use the same parameter name.

According to the postgres jdbc driver documentation:
"The timeout value used for socket read operations. If reading from the
server takes longer than this value, the connection is closed. This can
be used as both a brute force global query timeout and a method of
detecting network problems. The timeout is specified in seconds and a
value of zero means that it is disabled."