Support for @BeanParam when calling other services
This adds support for the @BeanParam when making http calls to external services using the HttpClient.

Since the HttpClient is not yet rewritten to make all reflection at startup, like the JaxRsResource-classes, I resorted to a simple cache to avoid doing the heavy reflection parts on each request.

Also added the Tuple class, feel free to use it wherever you like :-)