Add a customization point to jaxrs resource handling
JaxRsResourceInterceptor is an interface that allows for pluggable
interceptors that applications can register to customize calls to
JaxRsResources. Interceptors will be called before and after the call
to JaxRsResource.call(..)

Move JaxRsRequestContext into its own module (jaxrs-context).
The module registers a JaxRsResourceInterceptor that keeps a
context open during the resource call.