Fix support for hot reload of dao-method annotations
When you recompile a class during a debug session, it will be hot reloaded and replace the existing class. However, dao instances that has already been injected will still be using the old version of the interface and queries will therefore not update. This PR fixes that bug and allows for hot reloading to work again.