Upgrade logging
This will hopefully work out of the box with most existing configs.

If your config today contains:
```
layout: net.logstash.log4j.JSONEventLayoutV1
```

Then you would need to adjust to:
```
layout: JsonTemplateLayout
```
and also add the dependency:
```
<dependency>
   <groupId>org.apache.logging.log4j</groupId>
   <artifactId>log4j-layout-template-json</artifactId>
</dependency>
```

Support for file based appenders is not part of this PR but there I have some working code for that. Didn't want to feature creep in case it is not needed.