Added a nullcheck in the HttpClient for parameters annotated with a PathParam.
When passing a null value to a parameter annotated with PathParam in proxied resource method
a NullPointerException is thrown with an obscure message.

In this commit, we add a null check and handle the null by throwing an
IllegalArgumentException with a message that is more friendly and
improves the traceability.