Check minimum rows affected when executing statement in a batch
If statement is executed in a transaction then monoSink and fluxSink are not set for a statement.
If statement is also batched with another statement, then batch() and batchExecuted() methods are called instead of execute() method.
Current batchExecuted() method does nothing as fluxSink and monoSink are null, so ensureMinimumReached() method is never called.